+++ chaosmagpy/__init__.py
@@ -1,12 +1,18 @@
 __all__ = [
     "CHAOS", "load_CHAOS_matfile", "load_CHAOS_shcfile",
-    "load_CovObs_txtfile", "load_gufm1_txtfile",
+    "load_CovObs_txtfile", "load_gufm1_txtfile", "load_CALS7K_txtfile",
     "basicConfig", "synth_values"
 ]
 
 __version__ = "0.9-dev"
 
-from .chaos import (CHAOS, load_CHAOS_matfile, load_CHAOS_shcfile,
-                    load_CovObs_txtfile, load_gufm1_txtfile)
+from .chaos import (
+    CHAOS,
+    load_CHAOS_matfile,
+    load_CHAOS_shcfile,
+    load_CovObs_txtfile,
+    load_gufm1_txtfile,
+    load_CALS7K_txtfile
+)
 from .config_utils import basicConfig
 from .model_utils import synth_values